refactor: replace Error component with Notice for error handling
